agent/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-community
langchain-google-genai